Tilting Fix
Fixed the issue where the character sprite would tilt along with the turn by changing the target node for the click to a node above it and make another opositre tilt on the sprite to compensate the movement. Furthermore the line that caused the turning to push back to the initial orientation was deleted as this line was not needed for the functionality of the code and would cause steering issues.