Respawn system
Implements a system by wich when the player "falls" into a body of water they respawn on the last checkpoint they passed. This is done by taking the coordinates and orientation of the checkpoint via two points that are used to create the correct orientation so the player respawns facing the next segment of the track and avoid the player spawning and then driving backwards. Then the "respawn" is achieved by making the player position the checkpoint coordinates and making all movement and force zero so that the player does not mantain the momentum they had before the respawn.